requirements.txt
selenium>=4.22.0
webdriver-manager>=4.0.2
gspread>=6.0.0
google-auth>=2.30.0
python-dotenv>=1.0.0